fix: make the daily deck cover a component
