+++ scriptured-prayer-components/src/components/index.ts
@@ -8,3 +8,4 @@ export * from "./Footer";
 export * from "./FooterColumn";
 export * from "./FooterLinks";
 export * from "./FooterLink";
+export * from "./DailyDeck";
